requirements.txt
pandas
requests
matplotlib
mysql-connector-python
seaborn
python-dotenv